restructure rest folder
