+++ x/nft/client/rest/query.go
@@ -1,0 +1,32 @@
+package rest
+
+import (
+	"github.com/gorilla/mux"
+
+	"github.com/cosmos/cosmos-sdk/client/context"
+	"github.com/cosmos/cosmos-sdk/codec"
+)
+
+func registerQueryRoutes(cliCtx context.CLIContext, r *mux.Router,
+	cdc *codec.Codec, queryRoute string) {
+
+	// Get the total supply of a collection
+	r.HandleFunc(
+		"/nft/supply/{denom}", supplyNFTHandler(cdc, cliCtx, queryRoute),
+	).Methods("GET")
+
+	// Get the collections of NFTs owned by an address
+	r.HandleFunc(
+		"/nft/balance/{delegatorAddr}", getNFTsBalanceHandler(cdc, cliCtx, queryRoute),
+	).Methods("GET")
+
+	// Get all the NFT from a given collection
+	r.HandleFunc(
+		"/nft/{denom}", getCollectionHandler(cdc, cliCtx, queryRoute),
+	).Methods("GET")
+
+	// Query a single NFT
+	r.HandleFunc(
+		"/nft/{denom}/{id}", getNFTHandler(cdc, cliCtx, queryRoute),
+	).Methods("GET")
+}
